+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1]
+### :pencil: Changed
+- Add support for some ingredient volume convertion (Wine).
+
+### :nail: Polish
+- Refactor UNIT_TO_ML conversion table to recipeCalc/constants.ts
+
 ## [1.1.0]
 ### :pencil: Changed
 - Recipe YAML data-structure has been ditched in favor of markdown-like structure.